server_side/kuduq-consumer/requirements.txt
annotated-types==0.7.0
blessed==1.21.0
boto3==1.40.12
boto3-stubs==1.40.12
botocore==1.40.12
certifi==2025.8.3
chalice==1.32.0
charset-normalizer==3.4.2
click==8.2.1
dnspython==2.7.0
editor==1.6.6
email_validator==2.2.0
idna==3.10
inquirer==3.4.1
jmespath==1.0.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
readchar==4.2.1
requests==2.32.4
runs==1.2.2
s3transfer==0.13.1
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
wcwidth==0.2.13
xmod==1.8.1
